requirements/development.txt
-r base.txt
django-extensions==4.1
django-debug-toolbar